Update coins hacks
Hi, I had created a change yesterday but it was never approved. I was making sure that you understood that Blooket has SSC's (Server Side Checks). Meaning that it isn't possible to have the current value of coins and XP that are listed in the code. If you approve this small change, It will allow the code to run smoothly again. Thank You. 🐱‍👓